Backend/requirements.txt
# FastAPI Backend Requirements
# Python 3.10+

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.1.0

# Database (Async SQLAlchemy + SQLite/PostgreSQL)
sqlalchemy==2.0.35
aiosqlite==0.20.0
greenlet==3.1.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# AI Integration
groq==0.13.0

# CORS & Middleware
python-dotenv==1.0.1

# Production server
gunicorn==21.2.0
